# backend/docs/swagger.yaml
basePath: /
definitions:
  handlers.NotificationSettingsResponse:
    properties:
      daily_suggestion:
        type: boolean
      email_enabled:
        type: boolean
      monthly_summary:
        type: boolean
      push_enabled:
        type: boolean
      streak_reminder:
        type: boolean
      weekly_summary:
        type: boolean
    type: object
  handlers.SubscribePushRequest:
    properties:
      auth:
        type: string
      endpoint:
        type: string
      p256dh:
        type: string
      user_agent:
        type: string
    required:
    - auth
    - endpoint
    - p256dh
    type: object
...
  handlers.UnsubscribePushRequest:
    properties:
      endpoint:
        type: string
    required:
    - endpoint
    type: object
  handlers.UpdateNotificationSettingsRequest:
    properties:
      daily_suggestion:
        type: boolean
      email_enabled:
        type: boolean
      monthly_summary:
        type: boolean
      push_enabled:
        type: boolean
      streak_reminder:
        type: boolean
      weekly_summary:
        type: boolean
    type: object
  handlers.VAPIDKeyResponse:
    properties:
      vapid_public_key:
        type: string
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: RoambleのバックエンドAPIドキュメント
  title: Roamble API
  version: "1.0"
paths:
...
  /api/notifications/push/subscribe:
    delete:
      consumes:
      - application/json
      description: 指定endpointのPush通知購読をDBから削除する。存在しない場合も204を返す（冪等）。
      parameters:
      - description: 解除するendpoint
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.UnsubscribePushRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Push購読解除
      tags:
      - Notifications
    post:
      consumes:
      - application/json
      description: Push通知購読をDBに登録する。同一endpointが存在する場合はUpsert。
      parameters:
      - description: 購読情報
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.SubscribePushRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Push購読登録
      tags:
      - Notifications
  /api/notifications/push/vapid-key:
    get:
      description: Web Push通知用のVAPID公開鍵を返す
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.VAPIDKeyResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: VAPID公開鍵取得
      tags:
      - Notifications
  /api/notifications/settings:
    get:
      description: ユーザーの通知設定を取得する
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.NotificationSettingsResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 通知設定取得
      tags:
      - Notifications
    put:
      consumes:
      - application/json
      description: ユーザーの通知設定を更新する。送信されたフィールドのみ更新（部分更新）。
      parameters:
      - description: 更新する通知設定
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateNotificationSettingsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.NotificationSettingsResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 通知設定更新
      tags:
      - Notifications
  /api/places/{placeId}/photo:
    get:
      description: Place ID に対応する写真URLを取得する。キャッシュ未ヒット時のみ photo_reference を使って解決する
      parameters:
      - description: Place ID
        in: path
        name: placeId
        required: true
        type: string
      - description: Google Places Photo Reference（キャッシュ未ヒット時に必須）
        in: query
        name: photo_reference
        type: string
      - description: 画像最大幅ピクセル（デフォルト2000）
        in: query
        name: maxWidth
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 施設写真URL取得
      tags:
      - Places
...
securityDefinitions:
  BearerAuth:
    description: JWTアクセストークン（Bearer形式）
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"